[Refactor] Move wallet files to wallet dir
