+++ src/qt/fundamentalnodelist.cpp
@@ -9,7 +9,7 @@
 #include "fundamentalnodeconfig.h"
 #include "fundamentalnodeman.h"
 #include "sync.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 #include "walletmodel.h"
 
 #include <QMessageBox>
