+++ src/qt/masternodelist.cpp
@@ -9,7 +9,7 @@
 #include "masternodeconfig.h"
 #include "masternodeman.h"
 #include "sync.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 #include "walletmodel.h"
 
 #include <QMessageBox>
