+++ src/qt/multisigdialog.cpp
@@ -15,7 +15,7 @@
 #include "coins.h"
 #include "keystore.h"
 #include "init.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 #include "script/sign.h"
 #include "script/interpreter.h"
 #include "utilmoneystr.h"
