+++ src/qt/signverifymessagedialog.cpp
@@ -14,7 +14,7 @@
 
 #include "base58.h"
 #include "init.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 
 #include <string>
 #include <vector>
