+++ src/qt/splashscreen.cpp
@@ -15,7 +15,7 @@
 #include "version.h"
 
 #ifdef ENABLE_WALLET
-#include "wallet.h"
+#include "wallet/wallet.h"
 #endif
 
 #include <QApplication>
