+++ src/qt/vitae.cpp
@@ -37,7 +37,7 @@
 #include "util.h"
 
 #ifdef ENABLE_WALLET
-#include "wallet.h"
+#include "wallet/wallet.h"
 #endif
 
 #include <stdint.h>
