+++ src/test/multisig_tests.cpp
@@ -12,7 +12,7 @@
 #include "uint256.h"
 
 #ifdef ENABLE_WALLET
-#include "wallet_ismine.h"
+#include "wallet/wallet_ismine.h"
 #endif
 
 #include <boost/assign/std/vector.hpp>
