+++ src/test/zerocoin_denomination_tests.cpp
@@ -8,8 +8,8 @@
 #include "denomination_functions.h"
 #include "main.h"
 #include "txdb.h"
-#include "wallet.h"
-#include "walletdb.h"
+#include "wallet/wallet.h"
+#include "wallet/walletdb.h"
 #include <boost/test/unit_test.hpp>
 #include <iostream>
