+++ src/wallet/db.cpp
@@ -3,7 +3,7 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "db.h"
+#include "wallet/db.h"
 
 #include "addrman.h"
 #include "hash.h"
