+++ src/wallet/wallet_ismine.cpp
@@ -4,7 +4,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "wallet_ismine.h"
+#include "wallet/wallet_ismine.h"
 
 #include "key.h"
 #include "keystore.h"
