+++ src/wallet/walletdb.h
@@ -9,7 +9,7 @@
 #define BITCOIN_WALLETDB_H
 
 #include "amount.h"
-#include "db.h"
+#include "wallet/db.h"
 #include "key.h"
 #include "keystore.h"
 #include "primitives/zerocoin.h"
